+++ node/index.js
@@ -4,7 +4,12 @@ const config = {
     node: true
   },
   plugins: ['node'],
-  extends: ['plugin:node/recommended-module']
+  extends: ['plugin:node/recommended-module'],
+  settings: {
+    // Forbid the use of Node.js builtin modules
+    // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-nodejs-modules.md
+    'import/no-nodejs-modules': 'off'
+  }
 }
 
 module.exports = config